Surface declarable for headings and commodities
Jira link: https://transformuk.atlassian.net/browse/HOTT-811

We need to surface whether a search result is declarable (only ever true
for a commodity or heading) for the Single Trade Window team.

We also need to make sure that the STW team use the actual api endpoint
to do search so we're exposing this ability in the public api endpoint
whitelist configuration.