HOTT-988: Retry downloads of cf.deb file
We're getting invalid installs sometimes -
adding retry in the hope this is just flaky
network issues.

Oddly this is failing on the `dpkg -i` line and
not the `curl` line because the file is an
invalid deb. I'm guessing because its just an
empty file but I've added the `file cf.deb` line
so we can see what the files type is if/when it
fails to see if something else is going on.