HOTT-1417 Also send output to stdout
Previously we were not reporting the brakeman output to stdout.

The output from Brakeman, when there are no errors to show, doesn't fully comply with the JUnit format because there are no <testcase> nodes. Normally this doesn't matter because if there are errors, its valid, if there aren't, we don't need the output.

There is an additional <brakeman:properties> node but this does not match the JUnit XSD and is ignored by CircleCI.

When there is a parsing failure (as opposed to a scanning error), this is documented in the <brakeman:properties> section but this is ignore by CircleCI.

This results in parsing failures provided no information as to the source of the problem.