HOTT-1507 Switch to using the latest v1 ruby Orb
We're not running view specs in CI at present, this appears to be a bug
in the CircleCI Ruby orb. It's already been fixed upstream but we're not
benefitting because we're locked to a specific (old) version.

The Orb uses semver so switching `@1` allows us to benefit from bugfixes
whilst avoid any switches to incompatible versions.