HOTT-3165 Redefine max size of cookies to be lower
This works around a rails bug. Rails is computing the cookie size lower
than browsers. This means rails thinks the cookie is valid but browsers
are still rejecting them as being over 4kb

I _think_ this is caused by rails computing the cookie size pre-encoding
but browsers computing the size post encoding. This probably wasn't much
of a problem with signed cookies because there weren't many non ASCII
characters but with encrypted cookies the prevalance of characters
requiring encoding greatly increases - and the greater the occurrance of
these, the greater the skew in size calculation between server and
client