HOTT-3512: Remove VCAP... env var
This is a CloudFoundry specific environment variable. There's handling
code in lib/paas_config.rb to fall back to REDIS_URL for this purpose.

I've also moved the pre-commit config up one directory. It's scoped to
run hooks only on terraform/* for now so won't create huge diffs, but
it has to be in the root directory to run on the cli.