BAU: Fixes meursing
We added a new concept to our caching list

- View fragment caching

This threw me a bunch as I wasn't expecting to be investigating another
caching concept on top of CDN, redis fragments in the frontend and
backend fragments plus etags to boot.

TBH I'd probably like to just remove this as we really don't need to
cache these fragments.