HMRC-1843: Removes e2e tests during preview deployments
This is a cost/benefit choice based off of the slowness of these preview
environments and the fact that the timeout on the e2e suite got
increased to 60 second to accommodate it.

This timeout increase meant we didn't get the benefits of performance
checks that are implicit with each run every 15 minutes and the cost is
just too high

We've reset the timeout to be 10 seconds rather than 60 seconds so these
preview deployment tests will almost never pass.